+++ cipher-wrapper.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 cd /home/auctor/srv/ct-cipher
-exec ./dist/src/app/index.js "$@"
+exec node ./dist/src/app/index.js "$@"